requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn
h5py==2.10.0
itsdangerous==2.0.1
Jinja2==3.0.1
Keras==2.3.0
MarkupSafe==2.0.1
opencv-python==4.5.2.54
Pillow==7.0.0
six==1.16.0
scipy==1.4.1
tensorflow-cpu==2.2.0
Werkzeug==2.0.1
